refactor: migrate curve operations to gnark-crypto implementation
